Add Final Project for Android Dev. Class [First Iteration]
This project is made with flutter framework. From this, I learned the power of .gitignore. I already knew that when placed in the root directory of the repo, it could ignore all matching file. However, I wasn't aware that it could work as well in any other directories besides the root folder. In such case, the .gitignore file is local to that folder. Really a neat trick.